Some nice stuff
Battle almost sorted(?)

+++ AoSSimCPP/Die.h
@@ -1,4 +1,4 @@
 #pragma once
 #include <random>
 //Die Header
-uint8_t Roll();
+int Roll();
